+++ config.go
@@ -1,0 +1,10 @@
+package astm
+
+import (
+	"github.com/blutspende/go-astm/v3/models/astmmodels"
+)
+
+func NewDefaultConfiguration() astmmodels.Configuration {
+	config := astmmodels.DefaultConfiguration
+	return config
+}