+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.1.3] - 2025-06-16
+
+### Added
+
+### Changed
+- IdentifyMessage reworked
+
+### Fixed
+- Fixed multi message identification
+
 ## [3.1.2] - 2025-06-12
 
 ### Added